# Fix: Ensure iso_time uses user TZ when tz_override is empty